+++ examples/custom_checkers/e9995_type_is_assigned.py
@@ -1,14 +1,20 @@
-from typing import List, Dict
+from typing import List
+import datetime
 
-def add_two_numbers(x=int, y=List[float], z: type = complex) -> int: 
-    # type is assigned instead of annotated here, 
-    # should be def add_two_numbers(x: int, y: int) -> int
+
+class Person:
+    name = "Bob"
+
+
+def add_two_numbers(
+    x=int, # Error on this line
+    y=List[float], # Error on this line
+    z: type = complex # No error on this line
+) -> int:
     return (x + y) * z
 
 
 class MyDataType:
-    # type is assigned instead of annotated here
-    x = bool
-    y = Dict[str, str]
-    # checker should not pick this up
-    z: complex = complex
+    x = datetime.time # Error on this line
+    y = Person # Error on this line
+    z: complex = complex # No error on this line